refactoring vm_modules examples
